[component]: usbdevice: hid: fix GET_IDLE and SET_IDLE handlers
The USB HID standard declares mandatory support for GET_IDLE
and SET_IDLE requests for Boot Keyboard. Most hosts can handle
their absence, but others like some old/strange UEFIs and BIOSes
consider this a critical error and refuse to work with f_hid.

This primitive implementation of saving and returning idle is
sufficient to meet the requirements of the standard and these
devices.

And SET_IDLE value must be shifted 8 bits to the right to get
duration. This confirmed by USBCV test.

Change-Id: Icb83440bd57ba26bb5cbf5bdd1c84c474bc196dd
Signed-off-by: William Wu <william.wu@rock-chips.com>