[components]: usbdevice: msc: wait for disk ready
The usb msc function get the disk device during the
set config request of enumeration. For usb bus powered
products, the initialization of disk device may be
later than usb enumeration, this cause msc unable to
mount on Host.

Test on RK2108 board with 64GB EMMC and usb bus powered,
it is easy to encounter this problem.

This patch retry to get the disk device every 100ms
and timeout after wait for 5000ms.

Signed-off-by: William Wu <william.wu@rock-chips.com>
Change-Id: I25a36ab6301edd9cf005449944ab37a5b10f2c15