[BSP] Rockchip: Support clk gate from board file and gpio power mode
hclk might be different for different controllers, so need to
get it from board file if needed. And as we get clk_id from
board file already, we can use it for enabling instead of using
CLK_SDIO_GATE if defined.

Also we need to user gpio mode to enable power if needed.

Signed-off-by: Shawn Lin <shawn.lin@rock-chips.com>
Change-Id: I97ce0453080fb05d44d2a76cbdcc906b33a1989d