[BSP]: rockchip: common: don't retry transactions immediately after NAK
Current code when we get a NAK in bulk transfer, we immediately start
to retry transactions. This will cause some U-disk transfer error.

error log:
    CSW size error
    usb mass_storage write 2 sector failed

There is no requirement to retry super quickly after a NAK, so if we
received NAK, delay 30ms. We tested several U-disks and found that a
delay of 30ms can ensure stable reading and writing for all U-disks.

Change-Id: I4d8e1f6d5b0153807821b88ec5bf96a0a228b080
Signed-off-by: Jianwei Zheng <jianwei.zheng@rock-chips.com>