[BSP]: rockchip: common: fix USB Host occasionally unable to enumerate devices
On RK2118 platform we found that devices are still triggered to
connect after being plugged out. This will lead to the following
two issues.

1. The device has actually been disconnected, but due to the
trigger of the connect interrupt, the connect_status is set to
TRUE, The if condition will not be met when a device is plugged
in. Causing the device is unable to be enumerated.

2. Because the connect interrupt was triggered, the enumeration
process will be executed, but the device has disconnected,
resulting in the enumeration failure and causing the controller
to enter an abnormal state.

We found that after failing to enumerate the device in this
situation, The urb_state and hc_state will be set to idle state.
To resolve these two issues, we reinit the hcd and set the
connect_status to FALSE when urb_state and hc_state is in idle state.

Change-Id: Idc38078d047512ac3d022a4d67e2dd37f2997a78
Signed-off-by: Jianwei Zheng <jianwei.zheng@rock-chips.com>