[components] dfs: elmfat: Avoid fault when disk error
When the usb flash drive is suddenly unplugged in the process of writing
data, USB will automatically umount the USB flash drive, but the upper
application still holds the file descriptor of the USB flash drive, and
continues to find errors in the process of reading and writing. At this
time, it will release the semaphore of the file system, and this
semaphore has actually been released before, which leads to double free.

The crash backtrace like this:
0x1107a1a6
assert_hook
/home/rk/sdk/v4.1.x/components/utilities/cmbacktrace/cmb_port.c:269
0x11070e22
rt_mutex_release
/home/rk/sdk/v4.1.x/src/ipc.c:1103
0x11077cd4
f_write
/home/rk/sdk/v4.1.x/components/dfs/filesystems/elmfat/ff.c:3996
(discriminator 1)
0x1107590c
dfs_elm_write
/home/rk/sdk/v4.1.x/components/dfs/filesystems/elmfat/dfs_elm.c:582
0x11073712
write
/home/rk/sdk/v4.1.x/components/dfs/src/dfs_posix.c:186
0x1106eb18
usb_disk_rw_test
/home/rk/sdk/v4.1.x/bsp/rockchip/common/tests/usb_disk_rw_test.c:48

Change-Id: Ic09aea545e081767bd940c40e835265b78862f79
Signed-off-by: Cliff Chen <cliff.chen@rock-chips.com>