[components]: usbdevice: msc: improve performance for msc
This msc device driver read/write a block for block device
at a time. For nor flash, the size of a block is 4096 bytes.
For emmc, the size of a block is only 512 bytes. It's unable
to run the best performance if read/write a block at a time.

This patch adds a new config RT_USB_MSTORAGE_BUFLEN to set the
buffer size in each reading/writing for the block device. And
the default size is 4096 bytes. It also sets the size of usb
req the same as the value to get better IO transaction on usb
bus.

It must be noted that, this default value of RT_USB_MSTORAGE_BUFLEN
was decided experimentally as a trade-off between memory and
performance. Check the value of RT_USB_MSTORAGE_BUFLEN, if it
less than the size of a sector, return error immediately and
print log to help the user to reconfigure the correct value
of RT_USB_MSTORAGE_BUFLEN.

I have tested the performance of msc on RK2108 EVB and RK2108
recording pen board with ATTO Disk tool.

1. Without this patch
   - RK2108 EVB with nor flash
     read:  ~ 100 KBps
     write: ~ 921 6KBps

   - RK2108 recording pen board with FORSEE-NCEMAD9D-16G emmc
     read:  ~ 740 KBps
     write: ~ 1650 KBps

2. With this patch
   - RK2108 EVB with nor flash
     read:  ~ 130 KBps
     write: ~ 15400 KBps

   - RK2108 recording pen board with FORSEE-NCEMAD9D-16G emmc
     read:  ~ 3938 KBps
     write: ~ 3847 KBps

Change-Id: If37108578d9178d661fd7da5ba95fd7f7c101f96
Signed-off-by: William Wu <william.wu@rock-chips.com>